Clean up zeta methods
